% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarymethods.R
\name{coef.dcalasso}
\alias{coef.dcalasso}
\title{Extract coefficients from dcalasso objects}
\usage{
\method{coef}{dcalasso}(object, unpen = F, ...)
}
\arguments{
\item{object}{a dcalasso object}

\item{unpen}{whether to switch to the unpenalized coefficients}
}
\description{
\code{coef.dcalasso} extracts coefficients from dcalasso objects
}
\author{
Yan Wang, Tianxi Cai, Chuan Hong
}
