% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocess.R
\name{dataprocess}
\alias{dataprocess}
\title{Process data frame}
\usage{
dataprocess(mf, Terms, family)
}
\description{
\code{dataprocess} processes the original dataset into useful elements: Y, X, strata, weights, offset.
}
\author{
Yan Wang, Tianxi Cai, Chuan Hong
}
