% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcalasso-package.R
\docType{package}
\name{dcalasso-package}
\alias{dcalasso-package}
\title{Divide-and-Conquer adaptive lasso for Cox model with least square approximation}
\description{
\tabular{ll}{
Package: \tab dcalasso\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2018-04-09 \cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
Divide-and-conquer adaptive lasso for Cox model based on divide and conquer and least square approximation
}
\references{
the paper
}
\author{
Yan Wang  \email{yaw719@mail.harvard.edu}, Tianxi Cai \email{tcai@hsph.harvard.edu}, Chuan Hong <Chuan_Hong@hms.harvard.edu>
}
\keyword{approximation}
\keyword{divide-and-conquer}
\keyword{least}
\keyword{square}
