% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarymethods.R
\name{plot.dcalasso}
\alias{plot.dcalasso}
\title{Plot BIC paths for dcalasso objects}
\usage{
\method{plot}{dcalasso}(object, ...)
}
\arguments{
\item{object}{a dcalasso object}
}
\description{
\code{plot.dcalasso} summarizes output of dcalasso fit
}
\author{
Yan Wang, Tianxi Cai, Chuan Hong
}
