% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarymethods.R
\name{print.dcalasso}
\alias{print.dcalasso}
\title{Print dcalasso objects}
\usage{
\method{print}{dcalasso}(x, ...)
}
\arguments{
\item{x}{a dcalasso object}

\item{...}{...}
}
\description{
\code{print.dcalasso} summarizes output of dcalasso fit
}
\author{
Yan Wang, Tianxi Cai, Chuan Hong
}
