% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarymethods.R
\name{print.summary.dcalasso}
\alias{print.summary.dcalasso}
\title{Print summary for dcalasso objects}
\usage{
\method{print}{summary.dcalasso}(x, ...)
}
\arguments{
\item{x}{a summary.dcalasso object}

\item{...}{...}
}
\description{
\code{print.summary.dcalasso} summarizes output of dcalasso fit
}
\author{
Yan Wang, Tianxi Cai, Chuan Hong
}
