% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreinfofun.R
\name{scoreinfofun}
\alias{scoreinfofun}
\title{Computes score and information given subset}
\usage{
scoreinfofun(Y, X, stratas, weights, offset, family, bini, idx)
}
\arguments{
\item{Y}{outcome vector or matrix}

\item{X}{design matrix}

\item{stratas}{strata vector}

\item{weights}{weight for each observation}

\item{offset}{offset for each observation}

\item{family}{family of the outcome}

\item{bini}{initial coefficient estimate}

\item{idx}{indices for evaluation}
}
\description{
\code{scoreinfofun} computes information and scores.
}
\author{
Yan Wang, Tianxi Cai, Chuan Hong
}
