% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{SIM.FUN}
\alias{SIM.FUN}
\title{Generate simulation data to test adaptive lasso}
\usage{
SIM.FUN(nn, p.x = 50, cor = 0.2, family = c("binary", "count",
  "Cox"), beta0 = NULL)
}
\arguments{
\item{nn}{sample size}

\item{p.x}{number of covariates}

\item{cor}{correlation of covariates}

\item{family}{the family of response data taking c('binary','count','Cox')}

\item{beta0}{the coefficients for the design, including intercept}
}
\value{
For survival data, it returns a matrix with the first column U, second column delta (0,1), and rest = design matrix.
}
\description{
\code{SIM.FUN} generates continuous-time survival response data that are associated with design matrix.
The design matrix comes from a correlated multivariate normal. The default signals (beta0) are sparse.
}
\examples{
SIM.FUN(nn = 1e6, p.x = 50, family = 'binary')
}
\author{
Yan Wang, Tianxi, Chuan Hong
}
