% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarymethods.R
\name{summary.dcalasso}
\alias{summary.dcalasso}
\title{Summary method for dcalasso objects}
\usage{
\method{summary}{dcalasso}(object, unpen = F, ...)
}
\arguments{
\item{object}{a dcalasso fit}

\item{unpen}{whether to print out the unpenalized result}
}
\description{
\code{summary.dcalasso} summarizes output of dcalasso fit
}
\author{
Yan Wang, Tianxi Cai, Chuan Hong
}
