% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarymethods.R
\name{vcov.dcalasso}
\alias{vcov.dcalasso}
\title{Extract variance covariance from a dcalasso objects}
\usage{
\method{vcov}{dcalasso}(object, unpen = F, ...)
}
\arguments{
\item{object}{a dcalasso object}

\item{unpen}{whether to switch to the unpenalized variance covariance}
}
\description{
\code{vcov.dcalasso} extracts variance covariance objects
}
\author{
Yan Wang, Tianxi Cai, Chuan Hong
}
