% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wamstart.R
\name{warmfit}
\alias{warmfit}
\title{Warm-start fit wrapper}
\usage{
warmfit(Y, X, strata, weights, offset, family, idx)
}
\arguments{
\item{Y}{outcome vector or matrix}

\item{X}{design matrix}

\item{strata}{strata vector}

\item{weights}{weight for each observation}

\item{offset}{offset for each observation}

\item{family}{family of the outcome}

\item{idx}{indices for the subset to fit}
}
\description{
\code{warmfit} is to obtain an warm-start of the initial estimator using a subset of data.
}
\author{
Yan Wang, Tianxi Cai, Chuan Hong
}
