% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarymethods.R
\name{predict.dcalasso}
\alias{predict.dcalasso}
\title{Prediction of dcalasso object}
\usage{
\method{predict}{dcalasso}(object, newdata, type = "link")
}
\arguments{
\item{object}{a dcalasso object}

\item{newdata}{a new data frame}

\item{type}{"terms", "link", "response" same as predict.glm}
}
\description{
\code{predict.dcalasso} makes prediction of a dcalasso object based on the adaptive lasso estimation.
}
\author{
Yan Wang, Tianxi Cai, Chuan Hong
}
